Fix etpan deps creation
This was comparing a list with a string, which will start failing in
the future (as it should!).  But there's simply no need for this,
declare_dependency() function is quite happy with an empty list for
argument.  This was probably a leftover from an earlier attempt to get
things working.