Use features for openssl-camellia and openssl-idea
These actually switch on and off code, as opposed to the other HAVE
detction macros which only switch off header includes (and maybe
things would break if they're not there, but that's an existing bug
in netpgp).  So we want to be sure that they are present when we
need them.  This is what features are for.

This configures these features in delta-core as well.