Version the libdeltachat.so file
This creates a version scheme extracted from the existing git tagging
schema being used.  It therefore relies on this schema not changing.

Since there appear no way to do all of this reading things from git
before the project() call we can sadly not use the correct version
in the project itself.