Neppgp ssl mac (#347)
* Do not set HAVE_*_H defines for disabled features

It seems the mac build fails if we set the defines like this so
let's try avoiding including the header file alltogether.  I didn't
expect ssl to look at that define but there you go.

Maybe fixes #337

* Try to silence a few more warnings

Spotted on mac:
https://travis-ci.org/deltachat/deltachat-node/jobs/435493605

* Do not set HAVE_*_H defines for disabled features

It seems the mac build fails if we set the defines like this so
let's try avoiding including the header file alltogether.  I didn't
expect ssl to look at that define but there you go.

Maybe fixes #337

* Try to silence a few more warnings

Spotted on mac:
https://travis-ci.org/deltachat/deltachat-node/jobs/435493605

* booleans, how do they work