Update README.md
sudo -E ninja install

The -E parameter is necessary to get this commad working as expected.
Reason is because meson is installed for current user (pip) with given instructions and if we use sudo command
python3 is not able to find installed package. sudo sees another environment as current user.
The -E command passed current environment to sudo user.